refactor: move workflow.rs into workflow dir & extract validation mod into validation.rs
